--- a/src/basics.js
+++ b/src/basics.js
@@ -354,5 +354,6 @@ module.exports = {
   produce,
   subscribe,
   throwError,
+  transduce,
   transform
 };