--- a/src/basics.js
+++ b/src/basics.js
@@ -1,5 +1,5 @@
 // @flow
-const { Readable, Transform } = require('stream');
+const { Readable } = require('stream');
 const { StreamError } = require('./errors');
 const { wrapInPromise } = require('./utils');
 
@@ -9,6 +9,45 @@ type EventDone = {| done: true |};
 type EventError  = {| error: Error, done?: false |};
 type EventValue<T>  = {| value: T, done?: false |};
 export type Event<T> = EventDone | EventError | EventValue<T>;
+
+opaque type InternalStream<T> = stream.Readable;
+
+type SimpleProducer<ProducedT, SeedT> = (push: Push<ProducedT>, seed: ?SeedT) => ?SeedT;
+
+function createInternalStream<ProducedT, SeedT>(producer: SimpleProducer<ProducedT, SeedT> = (push) => {}, seed: ?SeedT): InternalStream<ProducedT> {
+  let currentSeed = seed;
+  return new Readable({
+    objectMode: true,
+    read() {
+      currentSeed = producer((event) => {
+        if (event.done) {
+          this.push(null);
+        }
+        else if (event.error) {
+          this.emit('error', event.error);
+        }
+        else {
+          // $FlowFixMe bug in the Readable type in flow, it does not support the object mode
+          this.push(event.value);
+        }
+      }, currentSeed);
+    }
+  });
+}
+
+function appendInternalStream<T>(stream: InternalStream<T>, event: Event<T>): InternalStream<T> {
+  if (event.done) {
+    stream.push(null);
+  }
+  else if (event.error) {
+    stream.emit('error', event.error);
+  }
+  else {
+    // $FlowFixMe bug in the Readable type in flow, it does not support the object mode
+    stream.push(event.value);
+  }
+  return stream;
+}
 
 opaque type StreamInternals<T> = {
   stream: Promise<stream.Readable>,
@@ -57,11 +96,13 @@ function transduce<T, TransformedT, AccumulationT>(
         }
       },
       end: () => {
+        // console.log('transduce listener end');
         accumulation = Promise.resolve(accumulation)
           .then((fulfilledAccumulation) => finalReducer(fulfilledAccumulation, { done: true }))
           .then(onFulfilled, onRejected);
       },
       error: (error) => {
+        // console.log('transduce listener error', error.toString());
         accumulation = Promise.resolve(accumulation)
           .then((fulfilledAccumulation) => finalReducer(fulfilledAccumulation, { error }))
           .then(onFulfilled, onRejected);
@@ -78,8 +119,8 @@ function transduce<T, TransformedT, AccumulationT>(
         const listener = createListener(readableStream, resolve, reject);
         readableStream
           .on('data', listener.data)
-          .on('end', listener.end)
           .on('error', listener.error)
+          .on('end', listener.end)
           .resume();
       }));
 
@@ -88,82 +129,35 @@ function transduce<T, TransformedT, AccumulationT>(
   };
 }
 
+function transduceToStream<ConsumedT, ProducedT>(transformer: ReducerTransformer<ConsumedT, ProducedT, any>): (Stream<ConsumedT>) => Stream<ProducedT> {
+  const transducer = transduce(transformer, appendInternalStream, createInternalStream);
+  return (stream) => wrapReadableStream(transducer(stream));
+}
+
 export type Push<T> = (Event<T>) => void;
-type PureTransformer<ConsumedT, ProducedT> = (event: Event<ConsumedT>, push: Push<ProducedT>) => Promise<void>;
 export type Transformer<ConsumedT, ProducedT, SeedT = void> = (event: Event<ConsumedT>, push: Push<ProducedT>, seed: ?SeedT) => ?SeedT | Promise<?SeedT>;
 
-function createTransformerFactory<ConsumedT, ProducedT, SeedT>(transformer: Transformer<ConsumedT, ProducedT, SeedT>, initialSeed: ?SeedT): () => PureTransformer<ConsumedT, ProducedT>{
-  const wrappedTransformer = wrapInPromise(transformer);
-  return () => {
-    let currentSeed = initialSeed;
-    return (event, push) => wrappedTransformer(event, push, currentSeed)
-      .then((newSeed) => {
-        currentSeed = newSeed;
-      });
-  };
-}
-
-function createTransformerPush<T>(stream: stream.Transform, { disableDone } = {}): Push<T> {
-  const onError = (event: EventError) => {
-    stream.emit('error', event.error);
-  };
-  const onDone = disableDone ?
-    (event: EventDone) => undefined :
-    (event: EventDone) => {
-      stream.emit('end');
-      stream.emit('finish');
-    };
-  const onValue = (event: EventValue<T>) => {
-    // $FlowFixMe bug in the Readable type in flow, it does not support the object mode
-    stream.push(event.value);
-  };
-  return (event: Event<T>) => {
-    if (event.error) {
-      onError(event);
-    }
-    else if (event.done) {
-      onDone(event);
-    }
-    else {
-      onValue(event);
-    }
-  };
-}
-
 function transform<ConsumedT, ProducedT, SeedT>(transformer: Transformer<ConsumedT, ProducedT, SeedT>, seed: ?SeedT): (Stream<ConsumedT>) => Stream<ProducedT> {
-  const transformerFactory = createTransformerFactory(transformer, seed);
-
-  return (stream: Stream<ConsumedT>) => {
-    if (stream.internals.consumer) {
-      throw new StreamError('Stream already being consumed.');
-    }
-    const wrappedTransformer = transformerFactory();
-    const transformer = new Transform({
-      objectMode: true,
-      transform(chunk, encoding, callback) {
-        wrappedTransformer({ value: chunk }, createTransformerPush(this))
-          .catch((error) => this.emit('error', error))
-          .then(callback);
-      },
-      flush(callback) {
-        wrappedTransformer({ done: true }, createTransformerPush(this, { disableDone: true }))
-          .catch((error) => this.emit('error', error))
-          .then(callback);
-      }
-    });
-    const transformedStream = wrapReadableStream(stream.internals.stream
-      .then((readableStream) => {
-        readableStream
-          .on('error', (error)  => {
-            wrappedTransformer({ error: error }, createTransformerPush(transformer))
-              .catch((error) => transformer.emit('error', error));
-          })
-          .pipe(transformer);
-        return transformer;
-      }));
-    stream.internals.consumer = transformedStream;
-    return transformedStream;
+  const reducerTransformer = (reducer) => {
+    const wrappedTransformer = wrapInPromise(transformer);
+    let currentSeed = seed;
+    return (accumulation, consumedEvent) => {
+      let accumulationPromise = Promise.resolve(accumulation);
+      return wrappedTransformer(
+        consumedEvent,
+        (producedEvent) => {
+          accumulationPromise = accumulationPromise
+            .then((accumulation) => reducer(accumulation, producedEvent));
+        },
+        currentSeed)
+        .then((newSeed) => {
+          currentSeed = newSeed;
+          return accumulationPromise;
+        });
+    };
   };
+
+  return transduceToStream(reducerTransformer);
 }
 
 type Subscriber<T> = (event: Event<T>) => Promise<void> | void;
@@ -189,100 +183,42 @@ function wrapReadableStream<T>(stream: stream.Readable | Promise<stream.Readable
   };
 }
 
-type ProducerP<ProducedT, SeedT> = (push: Push<ProducedT>, seed: ?SeedT) => Promise<?SeedT>;
-type Producer<ProducedT, SeedT> = (push: Push<ProducedT>, seed: ?SeedT) => Promise<?SeedT> | ?SeedT;
-
-function wrapProducer<ProducedT, SeedT>(producer: Producer<ProducedT, SeedT>, initialSeed: ?SeedT) {
-  const wrappedProducer: ProducerP<ProducedT, SeedT> = wrapInPromise(producer);
-  return (stream: Readable) => {
-    const theInitialSeed: ?SeedT = initialSeed;
-    return ({ buffer = [], productionEnded = false, seed = theInitialSeed } = {}) => {
-      // Even when paused one push is needed.
-      let expectPush = true;
-      // 1 - Let's deal with remaining events
-      while (buffer.length > 0 && expectPush) {
-        const event = buffer.shift();
-        if (event.error) {
-          stream.emit('error', event.error);
-          return { buffer: [], productionEnded, seed: (undefined: ?SeedT) };
-        }
-        else if (event.done) {
-          stream.push(null);
-          return { buffer: [], productionEnded, seed: (undefined: ?SeedT) };
-        }
-        else {
-          // $FlowFixMe bug in the Readable type in flow, it does not support the object mode
-          expectPush = stream.push(event.value);
-        }
-      }
-      // 2 - And now the new ones
-      if (productionEnded) {
-        return { buffer, productionEnded, seed };
-      }
-      return wrappedProducer(
-        (event: Event<ProducedT>) => {
-          if (productionEnded) {
-            throw new StreamError('No event should be produced once the stream has ended.');
-          }
-          else if (event.error) {
-            if (!expectPush) {
-              buffer.push(event);
-            }
-            else {
-              stream.emit('error', event.error);
-            }
-            productionEnded = true;
-          }
-          else if (event.done) {
-            if (!expectPush) {
-              buffer.push(event);
-            }
-            else {
-              stream.push(null);
-            }
-            productionEnded = true;
-          }
-          else {
-            if (!expectPush) {
-              buffer.push(event);
-            }
-            else {
-              // $FlowFixMe bug in the Readable type in flow, it does not support the object mode
-              expectPush = stream.push(event.value);
-            }
-          }
-        // $FlowFixMe not sure why this doesn't work here...
-        }, seed)
-        .then((seed: ?SeedT) => ({ buffer, productionEnded, seed }))
-        .catch((error) => {
-          if (productionEnded) {
-            throw error;
-          }
-          else if (expectPush) {
-            stream.emit('error', error);
-            return { buffer: [], productionEnded: true, seed: (undefined: ?SeedT) };
-          }
-          else {
-            buffer.push({ error });
-            return { buffer, productionEnded: true, seed: (undefined: ?SeedT) };
-          }
-        });
-    };
-  };
-}
+export type Producer<ProducedT, SeedT> = (push: Push<ProducedT>, seed: ?SeedT) => ?SeedT | Promise<?SeedT>;
 
 function produce<ProducedT, SeedT>(producer: Producer<ProducedT, SeedT>, seed: ?SeedT): Stream<ProducedT> {
-  const wrappedProducer = wrapProducer(producer, seed);
-  // A promise to make sure we don't push stuff out of order.
-  let producerPromise = Promise.resolve();
+  return wrapReadableStream(createInternalStream(
+    (push, currentSeed) => {
+      if (currentSeed instanceof Promise) {
+        return currentSeed.then((seed) => producer(push, seed));
+      }
+      else {
+        currentSeed = producer(push, currentSeed);
+        return currentSeed;
+      }
+    },
+    seed
+  ));
+}
 
-  const stream = new Readable({
-    objectMode: true,
-    read(size) {
-      producerPromise = producerPromise.then(wrappedProducer(this));
-    }
-  });
-  return wrapReadableStream(stream);
+function infinite<T>(): Stream<T> {
+  return produce((push) => {});
+}
+
+function append<T>(stream: Stream<T>, event: Event<T>): Stream<T> {
+  stream.internals.stream
+    .then((readableStream) => {
+      if (event.done) {
+        readableStream.push(null);
+      }
+      else if (event.error) {
+        readableStream.emit('error', event.error);
+      }
+      else {
+        // $FlowFixMe bug in the Readable type in flow, it does not support the object mode
+        readableStream.push(event.value);
+      }
+    });
+  return stream;
 }
 
 function isIterable(obj) {
@@ -295,37 +231,43 @@ function isIterable(obj) {
 
 function from<T>(input: Iterable<T> | stream.Readable): Stream<T> {
   if (input instanceof Readable) {
-    const transformer = new Transform({
-      objectMode: true,
-      transform(chunk, encoding, callback) {
-        this.push(chunk, encoding);
-        callback();
-      }
-    });
-    const transformedStream = wrapReadableStream(transformer);
-    // Forward errors to the transformed stream
-    input.on('error', (error) => transformer.emit('error', error));
-    input.pipe(transformer);
-    return transformedStream;
+    let stream = infinite();
+    const dataListener = (value) => append(stream, { value });
+    const errorListener = (error) => {
+      append(stream, { error });
+      input
+        .removeListener('data', dataListener)
+        .removeListener('error', errorListener)
+        .removeListener('end', endListener);
+    };
+    const endListener = () => {
+      append(stream, { done: true });
+      input
+        .removeListener('data', dataListener)
+        .removeListener('error', errorListener)
+        .removeListener('end', endListener);
+    };
+    input
+      .on('data', dataListener)
+      .on('error', errorListener)
+      .on('end', endListener)
+      .resume();
+    return stream;
   }
   else if (isIterable(input)) {
     // $FlowFixMe bug in the Iterable type of flow (cf. https://github.com/facebook/flow/issues/1163)
     const iterator: Iterator<T> = input[Symbol.iterator]();
-    return wrapReadableStream(new Readable({
-      objectMode: true,
-      read() {
-        for (;;) {
-          const { done, value } = iterator.next();
-          if (done) {
-            this.push(null);
-            return;
-          }
-          if (!this.push(value)) {
-            return;
-          }
+    return wrapReadableStream(createInternalStream(
+      (push) => {
+        const { done, value } = iterator.next();
+        if (done) {
+          push({ done: true });
+        }
+        else {
+          push({ value });
         }
       }
-    }));
+    ));
   }
   else {
     return throwError(new StreamError('Unable to create a stream, \'from\' only supports iterable or Readable stream.'));
@@ -337,12 +279,9 @@ function of<T>(...args: T[]): Stream<T> {
 }
 
 function throwError<T>(error: Error): Stream<T> {
-  return wrapReadableStream(new Readable({
-    objectMode: true,
-    read() {
-      this.emit('error', error);
-    }
-  }));
+  return produce((push) => {
+    push({ error });
+  });
 }
 
 module.exports = {
@@ -352,5 +291,6 @@ module.exports = {
   subscribe,
   throwError,
   transduce,
+  transduceToStream,
   transform
 };