… benchmark/drain.benchmark.js
… benchmark/helpers.js
… benchmark/mapReduce.benchmark.js
… package.json
… src/drain.js
… src/index.js
… src/internalStream.js
… src/map.js
… src/reduce.js
… src/tap.js
… test/drain.test.js
… test/map.test.js
… test/produce.test.js
… test/reduce.test.js
… test/subscribe.test.js
… test/throwError.test.js
… test/transform.test.js
… yarn.lock