Objective: getting performances on par with Node.JS stream.
---
_On f638064_ (introduction of the tests, before the work on this PR)

```console
$ NODE_ENV=production node ./benchmark/mapReduce.benchmark.js
alpes x 6.95 ops/sec ±3.36% (37 runs sampled)
node streams x 102 ops/sec ±1.74% (76 runs sampled)
highland x 471 ops/sec ±1.69% (79 runs sampled)
most x 23,633 ops/sec ±0.84% (82 runs sampled)
$ NODE_ENV=production node ./benchmark/drain.benchmark.js
alpes x 27.71 ops/sec ±4.40% (65 runs sampled)
node streams x 266 ops/sec ±4.18% (78 runs sampled)
highland x 4,374 ops/sec ±5.03% (64 runs sampled)
most x 47,610 ops/sec ±1.63% (79 runs sampled)
```

_On 133040c_ (after the work on this PR)

```console
$ NODE_ENV=production node ./benchmark/mapReduce.benchmark.js
alpes (map and reduce) x 72.02 ops/sec ±1.68% (79 runs sampled)
alpes (transduce) x 109 ops/sec ±3.67% (70 runs sampled)
node streams x 108 ops/sec ±1.78% (80 runs sampled)
highland x 517 ops/sec ±2.09% (82 runs sampled)
most x 14,833 ops/sec ±1.37% (80 runs sampled)
$ NODE_ENV=production node ./benchmark/drain.benchmark.js
alpes x 130 ops/sec ±1.03% (77 runs sampled)
node streams x 267 ops/sec ±1.44% (78 runs sampled)
highland x 4,263 ops/sec ±6.44% (55 runs sampled)
most x 48,291 ops/sec ±0.65% (80 runs sampled)
```

**Between x5 and x10 in terms on raw performances**