@cloderic 2018-02-26
Not perfect yet but good enough IMO