@cloderic 2018-02-27
Remarks taken into accounts @alexisfontaine !